+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All Notable changes to `bakame/tokei` will be documented in this file.
 ### Added
 
 - `IntervalSet::each`
+- `Duration::fromDateInterval`
 
 ### Fixed